flask_auth/requirements.txt
Flask
PyJWT
psycopg2-binary
bcrypt